「管理者チェックを追加」

+++ src/main/java/actions/EmployeeAction.java
@@ -24,7 +24,6 @@ public class EmployeeAction extends ActionBase {
 
         service = new EmployeeService();
 
-        //メソッドを実行
         invoke();
 
         service.close();
@@ -36,45 +35,42 @@ public class EmployeeAction extends ActionBase {
      * @throws IOException
      */
     public void index() throws ServletException, IOException {
+        if (checkAdmin()) {
 
-        //指定されたページ数の一覧画面に表示するデータを取得
         int page = getPage();
         List<EmployeeView> employees = service.getPerPage(page);
 
-        //全ての従業員データの件数を取得
         long employeeCount = service.countAll();
 
-        putRequestScope(AttributeConst.EMPLOYEES, employees); //取得した従業員データ
-        putRequestScope(AttributeConst.EMP_COUNT, employeeCount); //全ての従業員データの件数
-        putRequestScope(AttributeConst.PAGE, page); //ページ数
-        putRequestScope(AttributeConst.MAX_ROW, JpaConst.ROW_PER_PAGE); //1ページに表示するレコードの数
+        putRequestScope(AttributeConst.EMPLOYEES, employees);
+        putRequestScope(AttributeConst.EMP_COUNT, employeeCount);
+        putRequestScope(AttributeConst.PAGE, page);
+        putRequestScope(AttributeConst.MAX_ROW, JpaConst.ROW_PER_PAGE);
 
-        //セッションにフラッシュメッセージが設定されている場合はリクエストスコープに移し替え、セッションからは削除する
         String flush = getSessionScope(AttributeConst.FLUSH);
         if (flush != null) {
             putRequestScope(AttributeConst.FLUSH, flush);
             removeSessionScope(AttributeConst.FLUSH);
         }
 
-        //一覧画面を表示
         forward(ForwardConst.FW_EMP_INDEX);
+        }
 
     }
     public void entryNew() throws ServletException, IOException {
+        if (checkAdmin()) {
 
-        putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
-        putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView()); //空の従業員インスタンス
+        putRequestScope(AttributeConst.TOKEN, getTokenId());
+        putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView());
 
-        //新規登録画面を表示
         forward(ForwardConst.FW_EMP_NEW);
+        }
     }
 
 public void create() throws ServletException, IOException {
 
-    //CSRF対策 tokenのチェック
-    if (checkToken()) {
+    if (checkAdmin() && checkToken()) {
 
-        //パラメータの値を元に従業員情報のインスタンスを作成する
         EmployeeView ev = new EmployeeView(
                 null,
                 getRequestParam(AttributeConst.EMP_CODE),
@@ -85,76 +81,67 @@ public void create() throws ServletException, IOException {
                 null,
                 AttributeConst.DEL_FLAG_FALSE.getIntegerValue());
 
-        //アプリケーションスコープからpepper文字列を取得
         String pepper = getContextScope(PropertyConst.PEPPER);
 
-        //従業員情報登録
+
         List<String> errors = service.create(ev, pepper);
 
         if (errors.size() > 0) {
-            //登録中にエラーがあった場合
 
-            putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
-            putRequestScope(AttributeConst.EMPLOYEE, ev); //入力された従業員情報
-            putRequestScope(AttributeConst.ERR, errors); //エラーのリスト
+            putRequestScope(AttributeConst.TOKEN, getTokenId());
+            putRequestScope(AttributeConst.EMPLOYEE, ev);
+            putRequestScope(AttributeConst.ERR, errors);
 
-            //新規登録画面を再表示
             forward(ForwardConst.FW_EMP_NEW);
 
         } else {
-            //登録中にエラーがなかった場合
 
-            //セッションに登録完了のフラッシュメッセージを設定
             putSessionScope(AttributeConst.FLUSH, MessageConst.I_REGISTERED.getMessage());
 
-            //一覧画面にリダイレクト
             redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
         }
     }
 }
         public void show() throws ServletException, IOException {
+            if (checkAdmin()) {
 
-            //idを条件に従業員データを取得する
             EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
             if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
 
-                //データが取得できなかった、または論理削除されている場合はエラー画面を表示
                 forward(ForwardConst.FW_ERR_UNKNOWN);
                 return;
             }
 
-            putRequestScope(AttributeConst.EMPLOYEE, ev); //取得した従業員情報
+            putRequestScope(AttributeConst.EMPLOYEE, ev);
 
-            //詳細画面を表示
             forward(ForwardConst.FW_EMP_SHOW);
+            }
         }
 
         public void edit() throws ServletException, IOException {
+            if (checkAdmin()) {
 
-            //idを条件に従業員データを取得する
+
             EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
             if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
 
-                //データが取得できなかった、または論理削除されている場合はエラー画面を表示
                 forward(ForwardConst.FW_ERR_UNKNOWN);
                 return;
             }
 
-            putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
-            putRequestScope(AttributeConst.EMPLOYEE, ev); //取得した従業員情報
+            putRequestScope(AttributeConst.TOKEN, getTokenId());
+            putRequestScope(AttributeConst.EMPLOYEE, ev);
 
-            //編集画面を表示する
             forward(ForwardConst.FW_EMP_EDIT);
+            }
 
         }
 
         public void update() throws ServletException, IOException {
 
-            //CSRF対策 tokenのチェック
-            if (checkToken()) {
-                //パラメータの値を元に従業員情報のインスタンスを作成する
+            if (checkAdmin() && checkToken()) {
                 EmployeeView ev = new EmployeeView(
                         toNumber(getRequestParam(AttributeConst.EMP_ID)),
                         getRequestParam(AttributeConst.EMP_CODE),
@@ -165,28 +152,21 @@ public void create() throws ServletException, IOException {
                         null,
                         AttributeConst.DEL_FLAG_FALSE.getIntegerValue());
 
-                //アプリケーションスコープからpepper文字列を取得
                 String pepper = getContextScope(PropertyConst.PEPPER);
 
-                //従業員情報更新
                 List<String> errors = service.update(ev, pepper);
 
                 if (errors.size() > 0) {
-                    //更新中にエラーが発生した場合
 
-                    putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
-                    putRequestScope(AttributeConst.EMPLOYEE, ev); //入力された従業員情報
-                    putRequestScope(AttributeConst.ERR, errors); //エラーのリスト
+                    putRequestScope(AttributeConst.TOKEN, getTokenId());
+                    putRequestScope(AttributeConst.EMPLOYEE, ev);
+                    putRequestScope(AttributeConst.ERR, errors);
 
-                    //編集画面を再表示
                     forward(ForwardConst.FW_EMP_EDIT);
                 } else {
-                    //更新中にエラーがなかった場合
 
-                    //セッションに更新完了のフラッシュメッセージを設定
                     putSessionScope(AttributeConst.FLUSH, MessageConst.I_UPDATED.getMessage());
 
-                    //一覧画面にリダイレクト
                     redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
                 }
             }
@@ -194,17 +174,28 @@ public void create() throws ServletException, IOException {
 
         public void destroy() throws ServletException, IOException {
 
-            //CSRF対策 tokenのチェック
-            if (checkToken()) {
+            if (checkAdmin() && checkToken()) {
 
-                //idを条件に従業員データを論理削除する
                 service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
-                //セッションに削除完了のフラッシュメッセージを設定
                 putSessionScope(AttributeConst.FLUSH, MessageConst.I_DELETED.getMessage());
 
-                //一覧画面にリダイレクト
                 redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
             }
+        }
+        private boolean checkAdmin() throws ServletException, IOException {
+
+            EmployeeView ev = (EmployeeView) getSessionScope(AttributeConst.LOGIN_EMP);
+
+            if (ev.getAdminFlag() != AttributeConst.ROLE_ADMIN.getIntegerValue()) {
+
+                forward(ForwardConst.FW_ERR_UNKNOWN);
+                return false;
+
+            } else {
+
+                return true;
+            }
+
         }
     }
